add option to preserve calls to free in asmjs/wasm
The GDA detects if there are any calls to malloc, and if not, it will
remove any calls to free. This, however, doesn't work with
AddressSanitizer, which has defined its own malloc (which is not named
"malloc"). When compiled with Asan, the GDA would think that there are
no calls to malloc, when this is not actually the case.

This change adds an option that disables this free removal, which is
automatically enabled when compiling with sanitizers.

+++ llvm/lib/CheerpUtils/GlobalDepsAnalyzer.cpp
@@ -45,7 +45,7 @@ GlobalDepsAnalyzer::GlobalDepsAnalyzer(MATH_MODE mathMode_, bool llcPass)
 	  entryPoint(NULL), hasCreateClosureUsers(false), hasVAArgs(false),
 	  hasPointerArrays(false), hasAsmJSCode(false), hasAsmJSMemory(false), hasAsmJSMalloc(false),
 	  hasCheerpException(false), mayNeedAsmJSFree(false), llcPass(llcPass),
-	  hasUndefinedSymbolErrors(false), forceTypedArrays(false)
+	  hasUndefinedSymbolErrors(false), forceTypedArrays(false), preserveFree(PreserveFree)
 {
 }
 static void createNullptrFunction(llvm::Module& module)
@@ -694,7 +694,7 @@ bool GlobalDepsAnalyzer::runOnModule( llvm::Module & module )
 		Function* ffree = module.getFunction("free");
 		if (ffree)
 		{
-			if(!hasAsmJSMalloc)
+			if(!hasAsmJSMalloc && !preserveFree)
 			{
 				// The symbol is still used around, so keep it but make it empty
 				ffree->deleteBody();
